+++ RecipeApi/Models/IRecipe.cs
@@ -2,8 +2,8 @@
 {
 	public interface IRecipe
 	{
-		long Id { get; set; }
-		long? Category { get; set; }
+		int Id { get; set; }
+		int? Category { get; set; }
 		string? Name { get; set; }
 		string? Image { get; set; }
 		string? Description { get; set; }
