+++ RecipeApi/Models/SearchRecipeByTextRequest.cs
@@ -3,7 +3,7 @@
 	public class SearchRecipeByTextRequest
 	{
 		public string? SearchString { get; set; }
-		public long? CategoryId { get; set; }
+		public int? CategoryId { get; set; }
 		public int PageNumber {  get; set; }
 		public int RecordsPerPage { get; set; }
 	}
